requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
drf-yasg==1.21.10
git-filter-repo==2.47.0
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
whitenoise==6.9.0
